Delete unnecessary cp from make build command.
We had a cp -R -L command in our "make build" rule,
to convert symlinks to real files to support downstream
Python packaging steps, but those steps already
perform their own cp's that "materialize" symlinks,
so it's not necessary in the build step.

Note: Even if it weren't for the downstream Python
already materializing symlinks, it'd still be cleaner
for it to add support as necessary as opposed to
putting that step in the build rule that otherwise
has no need for it.